UPDATE: back

+++ packages/server/src/feature/organization/organization.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { OrganizationRepository } from './organization.repository';
 import { OrganizationMemberRepository } from './organization.member.repository';
 import { IOrganization, IOrganizationCreate, IOrganizationResponse, IOrganizationMemberResponse } from '@scspace-depot/types/organization';
@@ -13,13 +13,31 @@ export class OrganizationService {
   ) {}
 
   async createOrganization(organization: IOrganizationCreate): Promise<IOrganization> {
+
     const newOrganization = await this.organizationRepository.insert(organization);
     const newOrganizationMember = await this.organizationMemberRepository.insert(newOrganization.id, organization.delegatorId);
+
     return newOrganization;
   }
 
   async getOrganizationsByUserId(userId: number): Promise<IOrganization[]> {
-    return await this.organizationRepository.fetch(userId);
+    const organizations = await this.organizationRepository.fetch(userId);
+    if (!organizations) {
+      throw new NotFoundException('Organizations not found');
+    }
+    const organizationMembers = await this.organizationMemberRepository.find({ userId });
+    if (!organizationMembers) {
+      throw new NotFoundException('Organization members not found');
+    }
+    return organizations;
+  }
+
+  async deleteOrganization(organizationId: number): Promise<void> {
+    const organization = await this.organizationRepository.fetchById(organizationId);
+    if (!organization) {
+      throw new NotFoundException('Organization not found');
+    }
+    await this.organizationRepository.delete(organizationId);
   }
 
   async getOrganizations(): Promise<IOrganization[]> {
@@ -28,9 +46,11 @@ export class OrganizationService {
 
   async getOrganizationById(organizationId: number): Promise<IOrganizationResponse> {
     const organization = await this.organizationRepository.fetchById(organizationId);
+    if (!organization) {
+      throw new NotFoundException('Organization not found');
+    }
     const rawMembers = await this.organizationMemberRepository.find({ organizationId });
     const delegators = await this.userRepository.find({ id: organization.delegatorId });
-    
     // Get all member users
     const memberUsers = await this.userRepository.find({ 
       ids: rawMembers.map(member => member.userId) 
@@ -50,14 +70,39 @@ export class OrganizationService {
   }
 
   async getOrganizationMembers(organizationId: number): Promise<MOrganizationMember[]> {
+    const organization = await this.organizationRepository.fetchById(organizationId);
+    if (!organization) {
+      throw new NotFoundException('Organization not found');
+    }
     return await this.organizationMemberRepository.find({ organizationId });
   }
 
   async addMember(organizationId: number, userId: number): Promise<MOrganizationMember> {
+    const user = await this.userRepository.find({ id: userId });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    const organization = await this.organizationRepository.fetchById(organizationId);
+    if (!organization) {
+      throw new NotFoundException('Organization not found');
+    }
+    const organizationMember = await this.organizationMemberRepository.find({ organizationId, userId });
+    if (organizationMember.length > 0) {
+      throw new BadRequestException('User already in organization');
+    }
     return await this.organizationMemberRepository.insert(organizationId, userId);
   }
 
-  async removeMember(organizationId: number, userId: number): Promise<void> {
-    await this.organizationMemberRepository.delete(organizationId, userId);
+  async removeMember(organizationId: number, userId: number): Promise<boolean> {
+    const organization = await this.organizationRepository.fetchById(organizationId);
+    if (!organization) {
+      throw new NotFoundException('Organization not found');
+    }
+    const organizationMember = await this.organizationMemberRepository.find({ organizationId, userId });
+    if (!organizationMember) {
+      throw new NotFoundException('User not in organization');
+    }
+    const result = await this.organizationMemberRepository.delete(organizationId, userId);
+    return result;
   }
 }
